Feat: Added vulkan container and added support for dynamically creating objects by models using the entity manager.
The vulkan_manager is theoretically supposed to be a display_engine_manager where vulkan and directX can be used to send commands abstracted from the engine itself but that is for the future. The vulkan vertex buffers are now made by id and synced with entities inside the entity_manager. Also fixed the mouse jitteryness.