Update: Removed the event manager and made the player entities be an entity for uniformity.
The event manager was smited out of existance because I realised it is essentially the same as the command bus but with extra steps. This did mean I had to move player movement logic into the command bus. But the command bus needs cleaning up and restructuring anyway.

Player entities also now implement the entity trait. This is so they all fit inside a hashmap in the entitymanager for better uniformity.

BREAKING CHANGE!!
The dev command to delete the last created entity now also deletes the player when it gets to it since it's in the same map. This WILL crash the game.