Update: Updated entity_component to work with the new event_bus according to the observer pattern
Eventually I need to consider adding extra "fences" or somesuch so one
cant send an event to create an entity while the vulkan_component is
missing. Right now if that happens the whole thing just breaks or doesnt
do anything at all. Which *should* technically crash, but a developer
shouldnt really be allowed to do it on purpose.